move parameter path methods to ParameterPathConversions
