+++ packages/platform-core/src/index.ts
@@ -7,3 +7,4 @@ export * from './config/local_configuration_controller_factory.js';
 export * from './errors';
 export { USAGE_DATA_TRACKING_ENABLED } from './usage-data/constants.js';
 export { CDKContextKey } from './cdk_context_key.js';
+export * from './parameter_path_conversions.js';
